Bug-fixes for LogNVS: the measurement rate setting method was not working properly for meas rate > 1 Hz; the BINR protocol setting strategy by NMEA command was not switching the serial port to odd parity after receiving the acknowledgment

+++ src/main/java/org/gogpsproject/parser/nvs/NVSRateConfiguration.java
@@ -42,7 +42,7 @@ public class NVSRateConfiguration {
 		
 		msg.addElement(new Integer(nvsPrefix));
 		msg.addElement(new Integer(0xF4)); // F4h
-		msg.addElement(new Integer((int) ((1/measRate)/0.1)));
+		msg.addElement(new Integer((int) ((1/(double) measRate)/0.1)));
 		msg.addElement(new Integer(nvsPrefix));
 		msg.addElement(new Integer(nvsSuffix));
 	}
